pyproject.toml
[tool.poetry]
name = "repl-nix-cybershoptest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-login = "^0.6.3"
flask-sqlalchemy = "^3.1.1"
apscheduler = "^3.10.4"
gitpython = "^3.1.43"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
